singularity: Split up test code.
