+++ ext/singularity/testsingularity_test.go
@@ -1,0 +1,82 @@
+package singularity
+
+import (
+	"log"
+
+	"github.com/opentable/go-singularity/dtos"
+	"github.com/opentable/sous/lib"
+)
+
+// A testSingularity represents a test singularity instance.
+//
+// It provides functions that make it easy to construct a consistent
+// milieu in which tests can be run. The strategy for writing tests
+// with this is to construct a healthy and consistent world, and then
+// to introduce specific flaws against which tests can be written.
+type testSingularity struct {
+	Parent   *testFixture
+	BaseURL  string
+	Requests map[string]*testRequest
+}
+
+// AddCluster adds a cluster and ensures a singularity exists for its baseURL.
+// It creates the necessary singularity if it doesn't exist.
+//
+// It returns the singularity with the same base url.
+func (ts *testSingularity) AddCluster(name string) {
+	if ts.Parent.Clusters == nil {
+		ts.Parent.Clusters = sous.Clusters{}
+	}
+	cluster := &sous.Cluster{Name: name, BaseURL: ts.BaseURL}
+	ts.Parent.Clusters[name] = cluster
+}
+
+// AddRequest adds a new RequestParent and associated request to the test
+// fixture. The request parent created is identical to what defaultRequestParent
+// returns, except the ID is set to requestID. The configure func is passed this
+// request parent and may modify it before AddRequest returns it wrapped.
+//
+// It barfs if the requestID is not parseable with ParseRequestID, or if
+// it ends up with an empty Cluster or SourceLocation, or if the requestID is
+// not unique within this testSingularity, or if the cluster implied by the
+// request ID is not already defined.
+func (ts *testSingularity) AddRequest(requestID string, configure func(*dtos.SingularityRequestParent)) *testRequest {
+	deployID, err := ParseRequestID(requestID)
+	if err != nil {
+		log.Panicf("Error parsing requestID: %s", err)
+	}
+	if deployID.Cluster == "" {
+		log.Panicf("Request ID %q has an empty cluster component.", requestID)
+	}
+	if _, ok := ts.Parent.Clusters[deployID.Cluster]; !ok {
+		log.Panicf("Cluster %q not defined (from request id %q)", deployID.Cluster, requestID)
+	}
+	if deployID.ManifestID.Source.Repo == "" {
+		log.Panicf("Request ID %q has an empty source repo component.", requestID)
+	}
+
+	parent := &dtos.SingularityRequestParent{
+		RequestDeployState: &dtos.SingularityRequestDeployState{},
+		Request: &dtos.SingularityRequest{
+			Id:          requestID,
+			RequestType: dtos.SingularityRequestRequestTypeSERVICE,
+			Instances:   3,
+		},
+	}
+
+	if configure != nil {
+		configure(parent)
+	}
+	if ts.Requests == nil {
+		ts.Requests = map[string]*testRequest{}
+	}
+	request := &testRequest{
+		Parent:        ts,
+		RequestParent: parent,
+	}
+	if _, exists := ts.Requests[parent.Request.Id]; exists {
+		log.Panicf("request with ID %q already added", parent.Request.Id)
+	}
+	ts.Requests[parent.Request.Id] = request
+	return request
+}
